pyproject.toml
[project]
name = "mcp-github"
version = "0.1.0"
description = "Custom GitHub MCP server — kube-rbac-proxy fronts auth, shared GitHub App install behind."
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.18.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=43.0.0",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_github"]

[project.scripts]
mcp-github-stdio = "mcp_github.stdio:main"
mcp-github-http = "mcp_github.http:main"
